refactor: 행위에 따라 api, service 객체 분리
